% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftFunctions.R
\name{Char2Num}
\alias{Char2Num}
\title{Create character to numbers}
\usage{
Char2Num(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Given a dataframe, this function examines the class of each column; if a column's class is 'character', then it gets converted to a numeric type.
Vic's note: Not sure who wrote or used this function. Gets NAs if column cannot be stored as numeric.
}
\examples{
Char2Num()
}
\keyword{characters}
\keyword{numbers}
