% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhs-functions.R
\name{collate}
\alias{collate}
\title{Collate things}
\usage{
collate(x, y, sep = " ")
}
\description{
This function allows you to combine things together from different grids
}
\examples{
collate()

}
\keyword{collate}
