% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftFunctions.R
\name{efaEstimatedParameters}
\alias{efaEstimatedParameters}
\title{Number of parameters estimated in an EFA (oblique)}
\usage{
efaEstimatedParameters(items, numLatent)
}
\arguments{
\item{items}{numeric; number of items included in EFA}

\item{numLatent}{numeric; number of latent variables to be estimated}
}
\description{
Number of parameters estimated in an EFA (oblique)
}
