% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.bar.plot2.R
\name{likert.bar.plot2}
\alias{likert.bar.plot2}
\title{Modified Likert bar plot}
\usage{
likert.bar.plot2(l, low.color = "#D8B365", high.color = "#5AB4AC",
  neutral.color = "grey90", neutral.color.ramp = "white",
  colors = NULL, plot.percent.low = TRUE, plot.percent.high = TRUE,
  plot.percent.neutral = TRUE, plot.percents = FALSE, text.size = 3,
  text.color = "black", centered = TRUE, center = (l$nlevels - 1)/2 +
  1, include.center = TRUE, ordered = TRUE,
  wrap = ifelse(is.null(l$grouping), 50, 100), wrap.grouping = 50,
  legend = "Response", legend.position = "bottom",
  panel.arrange = "v", panel.strip.color = "#F0F0F0", group.order, ...)
}
\arguments{
\item{l}{likert class object}
}
\description{
This function is a modification of the likert.bar.plot() function in the likert package. Should provide items in the item order provided by dataframe columns. 
See sample script in 'R&D\_PRIVATE SHARED\Psychometric templates\Sample script for Likert Plots'
}
