% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_add_fullnames.df.R
\name{likert_add_fullnames.df}
\alias{likert_add_fullnames.df}
\title{Add names to likert graphs}
\usage{
likert_add_fullnames.df(rawDf, levels = 1:6, fnames)
}
\arguments{
\item{rawDf;}{a data frame with named columns}

\item{levels;}{a numeric vector indicating the number of levels per item}

\item{fnames;}{a character vector providing the item content names to be placed in the column names.}
}
\description{
Takes a list of dataframe column names, and long names, and stores the long names in the attribute 'fullname'. Then, the original names of the original dataframe get renamed with the full names.
This is to be used with the 'likert' package. e.g., WHen plotting, item names should appear with item content description if this function is used.
Example:
db <- add_likert_fullnames(db, c(
 'X7'='Do you use sites like Facebook, Twitter, or GPlus?',
 'X10'='Do you know of online communities relevant to your discipline or the courses you are taking now?'))
}
