% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.miss.R
\name{mean.miss}
\alias{mean.miss}
\title{Calculate mean with allowable missingness}
\usage{
\method{mean}{miss}(rowVector, maxMiss)
}
\arguments{
\item{rowVector}{vector; a row vector in, say, a data frame of Cases by Variables}

\item{maxMiss}{numeric; a number specifying the maximum NAs that are allowed prior to calculation}

\item{items}{numeric; number of cases to sample from each strata}
}
\description{
This function calculates a mean if the missingness does not exceed that maximum allowed.
}
\examples{
\dontrun{
#raw data
x <- c(1,2,2,3,4,1,2,NA, NA)
mean.miss(x, 1)
mean.miss(x, 2)
mean.miss(x, 2)
}
}
