% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.miss.R
\name{proSumScore}
\alias{proSumScore}
\title{Prorate subscale sum scores}
\usage{
proSumScore(rowVector, maxMiss)
}
\arguments{
\item{rowVector}{vector; a row vector in, say, a data frame of Cases by Variables. The row vector contains ONLY columns that contribute to the sum score.}

\item{maxMiss}{numeric; a number specifying the maximum NAs that are allowed prior to calculation}
}
\description{
This function computes a sum score across items in the event of missingness, but also prorates.
}
