% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round half up}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{number}

\item{digits}{numeric; a number specifying the number of digits to round to.}
}
\description{
R's default round() function does banker's rounding. This round2 function rounds a number half up.
}
