% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedNorm.R
\name{stratifiedNorm}
\alias{stratifiedNorm}
\title{Function to draw observations from stratified norm sample}
\usage{
stratifiedNorm(dat, strata, observations = 0, return.grid = FALSE,
  full.data = FALSE, full.data.id = "sampled")
}
\arguments{
\item{dat}{data.frame object}

\item{strata}{character; a named vector indicating the strata variables}

\item{observations}{numeric; number of cases to sample from each strata}

\item{return.grid}{logical; return grid of factor combintions?}

\item{full.data}{logical; return the original dataset with a new logical
variable (named from 'full.data.id') indicating whether the row was selected}

\item{full.data.id;}{a string pertaining the variable name used when 'full.data=TRUE'}
}
\description{
This function creates a stratified random sample, given an arbitrary number of
factors and levels within those factors. Credits to Matthew Sigal for finding and modifying this code.
}
