% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhs-functions.R
\name{Char2Num}
\alias{Char2Num}
\title{Create character to numbers}
\usage{
Char2Num(df)
}
\description{
This function allows you to change character text to numbers
}
\examples{
Char2Num()
}
\keyword{characters}
\keyword{numbers}
