% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efaEstimatedParameters.R
\name{efaEstimatedParameters}
\alias{efaEstimatedParameters}
\title{Number of parameters estimated in an EFA}
\usage{
efaEstimatedParameters(items, factors)
}
\arguments{
\item{items}{numeric; number of items included in EFA}

\item{factors}{numeric; number of latent variables to be estimated}
}
\description{
Number of parameters estimated in an EFA
}
