% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCliffsD.R
\name{getCliffsD}
\alias{getCliffsD}
\title{Cliff's d}
\usage{
getCliffsD(col, g)
}
\arguments{
\item{col}{a numeric vector with data values}

\item{g}{numeric; a number specifying the number of digits to round to.}
}
\description{
Wrapper function to call effsize::cliff.delta. Returns only the Cliff's delta estimate and its confidence intervals. For further details, see original function.
}
