% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proSumScore.R
\name{proSumScore}
\alias{proSumScore}
\title{Prorate subscale sum score}
\usage{
proSumScore(rowVector, maxMiss)
}
\arguments{
\item{rowVector}{vector; a numeric vector whose elements are to be summed and prorated}

\item{maxMiss}{numeric; a number specifying the maximum NAs that are allowed prior to calculation}
}
\description{
This function computes a sum score across items in the event of missingness, but also prorates.
}
\examples{

}
